Added WordsLayouterBase class
